Add authorization into services

+++ exam-service/src/main/java/com/ercanbeyen/examservice/controller/ExamController.java
@@ -2,6 +2,7 @@ package com.ercanbeyen.examservice.controller;
 
 import com.ercanbeyen.examservice.dto.ExamDto;
 import com.ercanbeyen.examservice.service.ExamService;
+import com.ercanbeyen.examservice.util.AuthUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,29 +14,35 @@ import java.util.List;
 @RequestMapping("/exams")
 public class ExamController {
     private final ExamService examService;
+    private final AuthUtil authUtil;
 
     @PostMapping
-    public ResponseEntity<ExamDto> createExam(@RequestBody ExamDto request) {
+    public ResponseEntity<ExamDto> createExam(@RequestBody ExamDto request, @RequestHeader("loggedInUser") String username) {
+        authUtil.checkUserHasAdminRole(username);
         return ResponseEntity.ok(examService.createExam(request));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ExamDto> updateExam(@PathVariable String id, @RequestBody ExamDto request) {
+    public ResponseEntity<ExamDto> updateExam(@PathVariable String id, @RequestBody ExamDto request, @RequestHeader("loggedInUser") String username) {
+        authUtil.checkUserHasAdminRole(username);
         return ResponseEntity.ok(examService.updateExam(id, request));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ExamDto> getExam(@PathVariable String id) {
+    public ResponseEntity<ExamDto> getExam(@PathVariable String id, @RequestHeader("loggedInUser") String username) {
+        authUtil.checkUserHasAdminRole(username);
         return ResponseEntity.ok(examService.getExam(id));
     }
 
     @GetMapping
-    public ResponseEntity<List<ExamDto>> getExams() {
+    public ResponseEntity<List<ExamDto>> getExams(@RequestHeader("loggedInUser") String username) {
+        authUtil.checkUserHasAdminRole(username);
         return ResponseEntity.ok(examService.getExams());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteExam(@PathVariable String id) {
+    public ResponseEntity<String> deleteExam(@PathVariable String id, @RequestHeader("loggedInUser") String username) {
+        authUtil.checkUserHasAdminRole(username);
         return ResponseEntity.ok(examService.deleteExam(id));
     }
 }
